refactor(core,tools): post-merge cleanup of the capability + coverage work
Quality pass over #130/#131/#132 (no behavior change — snapshots regenerate
byte-identical, check-parity --all stays at 290 OK).

- Move getActivitiesInWindow from distribution.ts to metric-input.ts, the
  home for shared metric accessors (alongside getActivities). Removes the
  capability.ts -> distribution.ts cross-tier dependency the durability port
  introduced; future metric tiers import the windowing helper from the
  accessor layer, not sideways from distribution.

- Extract the golden-fixture allowlist (HARNESS_FIXTURES + DEFAULT_FROZEN_NOW)
  into tools/harness-fixtures.ts as a single source of truth. The coverage
  probe now builds its per-fixture anchor manifest straight from that shared
  list instead of hand-mirroring the anchor table (the prior "keep in step"
  comment was a standing drift hazard) and no longer reads manifest.json.

- measure-reference-coverage-native.py: collapse the duplicated coverage
  analysis — one _analyze() pass now yields both the missing branch arcs and
  the branch totals (was two passes + analysis2); compute the entered-function
  set once instead of re-scanning it for the summary count.

- Drop a pre-existing unused readdirSync import from snapshot-section-11.ts
  (tools/ isn't in the CI lint set, so it was never flagged).
